fix: deep audit R2 — resource leaks, concurrency, error handling
- Fix TempDir use-after-free in e2e/support/mod.rs (server ran on deleted dir)
- Fix TempDir use-after-free in provider_events tests (2 files, latent bugs)
- Fix webhook fire-and-forget: two-spawn pattern with panic logging + permit safety
- Document OidcProvider std::sync::Mutex as correct (short, non-awaiting critical section)
- Add tracing::warn! for silent directory scan errors in config_io.rs
- Add tracing::debug! for LFS OID parsing errors (7 locations in lfs.rs)
- Fix AtomicBool ordering: Relaxed→Acquire/Release for shutdown flags
- Fix useless u64::from() conversions after u64 format migration
- Update packed_start/packed_end to accept u64 (consistent with format migration)
- 7,963 tests pass, CI green

+++ crates/fuzz/fuzz_targets/shardline/shard_parse.rs
@@ -89,7 +89,7 @@ fn summarize_shard(data: &[u8]) -> Result<ShardSummary, String> {
         }
         let file_size = file.file_size();
         let summed = file.segments.iter().try_fold(0_u64, |total, segment| {
-            total.checked_add(u64::from(segment.unpacked_segment_bytes))
+            total.checked_add(segment.unpacked_segment_bytes)
         });
         if summed != Some(file_size) {
             return Err("file segment lengths disagreed with file size".to_owned());
